% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute Missing Values in Omics Data}
\usage{
impute_missing(data, method = c("mean", "median"))
}
\arguments{
\item{data}{Omics data frame with missing values.}

\item{method}{Imputation method ("mean", "median").}
}
\value{
Data frame with imputed values.
}
\description{
Impute Missing Values in Omics Data
}
