% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_omics.R
\name{integrate_omics}
\alias{integrate_omics}
\title{Integrate Proteomic and Metabolomic Data}
\usage{
integrate_omics(prot, met)
}
\arguments{
\item{prot}{Proteomics data frame.}

\item{met}{Metabolomics data frame.}
}
\value{
A merged data frame with common samples.
}
\description{
Integrate Proteomic and Metabolomic Data
}
