% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_omics.R
\name{normalize_omics}
\alias{normalize_omics}
\title{Normalize Omics Data}
\usage{
normalize_omics(data, method = c("zscore", "log2"))
}
\arguments{
\item{data}{A numeric data frame of omics values.}

\item{method}{Normalization method ("zscore", "log2", "quantile").}
}
\value{
Normalized data frame.
}
\description{
Normalize Omics Data
}
